detect-default-branch-module-placement: move default branch detection to submission module
